+++ src/common/metrics/positive_rate.tsx
@@ -9,8 +9,8 @@ import { MetricDefinition } from './interfaces';
 import ExternalLink from 'components/ExternalLink';
 import Thermometer from 'components/Thermometer';
 import { metricToTooltipContentMap } from 'cms-content/infoTooltips';
-import StyledTooltip from 'components/InfoTooltip/StyledTooltip';
-import { StyledMarkdown } from 'components/InfoTooltip/InfoTooltip.style';
+import { InfoTooltip } from 'components/InfoTooltip';
+import { StyledMarkdown } from 'components/InfoTooltip/Tooltip.style';
 
 const METRIC_NAME = 'Positive test rate';
 
@@ -249,5 +249,5 @@ function renderInfoTooltip(): React.ReactElement {
   const tooltipContent = metricToTooltipContentMap[Metric.POSITIVE_TESTS];
   const { body } = tooltipContent;
 
-  return <StyledTooltip title={<StyledMarkdown source={body} />} />;
+  return <InfoTooltip title={<StyledMarkdown source={body} />} />;
 }
