+++ src/components/InfoTooltip/index.ts
@@ -1,0 +1,6 @@
+import InfoTooltip from './InfoTooltip';
+import { TooltipProps } from '@material-ui/core/Tooltip';
+
+export type StyledTooltipProps = Omit<TooltipProps, 'children'>;
+
+export { InfoTooltip };
